fix all register controller bugs

+++ src/utils/cloudinary.js
@@ -2,30 +2,40 @@ import {v2 as cloudinary} from "cloudinary"
 import fs from "fs"
 
 
-cloudinary.config({ 
-  cloud_name: process.env.CLOUDINARY_CLOUD_NAME, 
+cloudinary.config({
+  cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
   api_key: process.env.CLOUDINARY_API_KEY,
   api_secret: process.env.CLOUDINARY_API_SECRET,
 });
 
+console.log("Cloud Name:", process.env.CLOUDINARY_CLOUD_NAME);
+console.log("API Key:", process.env.CLOUDINARY_API_KEY);
+console.log("API Secret Exists:", !!process.env.CLOUDINARY_API_SECRET);
 
-const uploadOnCloudinary = async (localFilePath) => {
+
+
+const uploadToCloudinary = async (localFilePath) => {
     try {
         if(!localFilePath) return null
         //upload the file on cloudinary
         const response = await cloudinary.uploader.upload(localFilePath, {
-            resource_type: "auto"
+            resource_type: "auto",
         })
         // file has been uploaded successfull
-        console.log("file is uploaded on cloudinary", response.url);
+        //console.log("file is uploaded on cloudinary", response.url);
         return response;
+        fs.unlinkSync(localFilePath)
 
     } catch (error) {
-        fs.unlinkSync(localFilePath) // remove the locally saved temporary file as the upload operation got failed
-        return null;
+    console.error("Cloudinary Upload Error:", error);
 
+    if (localFilePath && fs.existsSync(localFilePath)) {
+        fs.unlinkSync(localFilePath);
     }
+
+    return null;
+}
 }
 
 
-export {uploadOnCloudinary}
+export {uploadToCloudinary}
